move structures to Structures directory and namespace
