+++ ChebsNecromancy/Structures/NeckroGathererPylon.cs
@@ -7,7 +7,7 @@ using Jotunn.Entities;
 using UnityEngine;
 using Logger = Jotunn.Logger;
 
-namespace ChebsNecromancy.Minions
+namespace ChebsNecromancy.Structures
 {
     internal class NeckroGathererPylon : MonoBehaviour
     {
